Rename IOutput to PsGraphOutputWriter refactor session initializer.
