+++ src/Authentication/Authentication.Core/Interfaces/IAuthContext.cs
@@ -4,7 +4,6 @@
 
 namespace Microsoft.Graph.PowerShell.Authentication
 {
-    using System;
     using System.Security.Cryptography.X509Certificates;
 
     public enum AuthenticationType
@@ -27,15 +26,6 @@ namespace Microsoft.Graph.PowerShell.Authentication
         DeviceCodeProviderFallBack,
         ClientCredentialProvider,
         UserProvidedToken
-    }
-
-    public interface IOutput
-    {
-        Action<string> WriteObject { get; set; }
-        Action<string> WriteDebug { get; set; }
-        Action<Exception, string, int, object> WriteError { get; set; }
-        Action<object, string[]> WriteInformation { get; set; }
-        Action<string> WriteVerbose { get; set; }
     }
 
     public interface IAuthContext
