+++ src/Authentication/Authentication.Core/Interfaces/IPSGraphOutputWriter.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Microsoft.Graph.PowerShell.Authentication
+{
+    public interface IPSGraphOutputWriter
+    {
+        Action<string> WriteObject { get; set; }
+        Action<string> WriteDebug { get; set; }
+        Action<Exception, string, int, object> WriteError { get; set; }
+        Action<object, string[]> WriteInformation { get; set; }
+        Action<string> WriteVerbose { get; set; }
+    }
+}
