+++ src/Authentication/Authentication/Models/PsGraphOutputWriter.cs
@@ -2,11 +2,11 @@
 //  Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
 // ------------------------------------------------------------------------------
 
-namespace Microsoft.Graph.PowerShell.Authentication.Common
+namespace Microsoft.Graph.PowerShell.Authentication.Models
 {
     using System;
 
-    public class Output : IOutput
+    internal class PsGraphOutputWriter : IPSGraphOutputWriter
     {
         public Action<string> WriteObject { get; set; }
         public Action<string> WriteDebug { get; set; }
